fixed issues and added the contact form
Issues were related to styles being overwritten and the navigation being
limited in size.
Added the contact form and styled it appropriately.
Wanted to automate making of the skills section, but I can't have a loop
inside of a function or loops nested. I'll try to come up with a better
workaround or just change the compiler since gulp-haml-coffee isn't
being maintained.